create util for tests
